GDB-9108 refactor cluster management module
## What
* Split controllers and services in separate files and packages.

## Why
* The reason is to improve the maintainability.

## How
* Split controllers and moved in separate files and all controllers moved in the controllers package.
* Moved services in separate package.
