+++ src/js/angular/clustermanagement/app.js
@@ -1,6 +1,6 @@
 import 'angular/core/services';
 import 'angular/core/directives';
-import 'angular/clustermanagement/controllers';
+import 'angular/clustermanagement/controllers/cluster-management.controller';
 import 'angular/clustermanagement/directives';
 import 'angular/core/services/repositories.service';
 import 'lib/d3.patch.js';
@@ -9,8 +9,8 @@ import 'angular-pageslide-directive/dist/angular-pageslide-directive';
 const modules = [
     'ngAnimate',
     'toastr',
-    'graphdb.framework.clustermanagement.controllers',
+    'graphdb.framework.clustermanagement.controllers.cluster-management',
     'graphdb.framework.clustermanagement.directives'
 ];
 
-const clusterManagementApp = angular.module('graphdb.framework.clustermanagement', modules);
+angular.module('graphdb.framework.clustermanagement', modules);
